chore: update eslint configuration
