+++ src/types/next.ts
@@ -1,30 +1,4 @@
-import type {
-  GetStaticProps,
-  InferGetStaticPropsType,
-  NextApiHandler,
-  NextComponentType,
-} from "next";
-import type { AppContextType, AppPropsType } from "next/dist/shared/lib/utils";
-import type { NextRouter } from "next/router";
-import type { Session } from "next-auth";
-
-export type ApiHandler<ResponseBody = unknown> = NextApiHandler<ResponseBody>;
-
-/**
- * TODO: Currently `router` type is inferred as `any`. Simplify this type
- * definition when the issue is resolved.
- *
- * @example export type Application = AppType<ApplicationPageProps>
- */
-export type Application = NextComponentType<
-  AppContextType,
-  ApplicationPageProps,
-  AppPropsType<NextRouter, ApplicationPageProps>
->;
-
-export type ApplicationPageProps = {
-  session: Session;
-};
+import type { GetStaticProps, InferGetStaticPropsType } from "next";
 
 export type InferGetStaticProps<GetStaticPropsFunction extends GetStaticProps> =
   Omit<InferGetStaticPropsType<GetStaticPropsFunction>, "trpcState">;
